WIP: Monitor mysql table stats
TODO this still needs to give zabbix SELECT permissions to the tables,
otherwise the table statistics are not available. Do we want zabbix to
be able to read everything? Not sure yet, leaning towards no.

I don't want to lose the other changes though so I'm commiting this here
for now.

Signed-off-by: Florian Pritz <bluewind@xinu.at>