Add prometheus node exporter for reproducible.archlinux.org
Extend the prometheus role to open up a firewall port for
monitoring.archlinux.org for the node-exporter port. In the future the
ip of monitoring.archlinux.org should not be hardcoded in ansible but
retrieved using ansible host variables.

The open firewalled ports cannot be removed out of the box by the
firewalld module and have to be an exact match of the old ip address.